requirements.txt
webull==0.6.1
python-dotenv==1.0.0
psutil>=7.0.0
requests>=2.25.0
pytz>=2021.1
